LPS-163461 Create Util file for Filter functions
